+++ src/app/dashboard/(registre-securite)/sites/[siteId]/_components/LotsTab.tsx
