+++ src/app/dashboard/(registre-securite)/sites/[siteId]/_components/PartsTab.tsx
